src/misc/requirements.txt
wordcloud==1.9.2
Pillow==9.5.0